+++ src/fastga/models/aerodynamics/components/__init__.py
@@ -11,7 +11,7 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .compute_cl_extreme import ComputeExtremeCL
+from .compute_cl_extreme import ComputeAircraftMaxCl
 from .compute_reynolds import ComputeUnitReynolds
 from .compute_cnbeta_fuselage import ComputeCnBetaFuselage
 from .compute_L_D_max import ComputeLDMax
@@ -29,3 +29,5 @@ from .compute_vn import ComputeVNAndVH, ComputeVN
 from .compute_cm_alpha_fus import ComputeFuselagePitchingMoment
 from .compute_equilibrated_polar import ComputeEquilibratedPolar
 from .compute_non_equilibrated_polar import ComputeNonEquilibratedPolar
+from .compute_cl_extreme_wing import ComputeExtremeCLWing
+from .compute_cl_extreme_htp import ComputeExtremeCLHtp
